requirements.txt
# Core
robotframework==6.2.3
robotframework-seleniumlibrary==6.1.1

# Selenium WebDriver
selenium==4.12.0
webdriver-manager==4.1.1

# Utilities
PyYAML==6.0
Faker==18.10.1
